+++ projects/plugins/jetpack/modules/blaze.php
@@ -17,5 +17,20 @@ use Automattic\Jetpack\Blaze;
 
 Blaze::init();
 
-// Remove post row Blaze actions in the Jetpack plugin.
-add_filter( 'jetpack_blaze_post_row_actions_enable', '__return_false' );
+/**
+ * Remove post row Blaze actions in the Jetpack plugin.
+ * Keep them on for products.
+ *
+ * @param bool    $are_quick_links_enabled Should Blaze row actions be enabled.
+ * @param WP_Post $post                    The current post in the post list table.
+ *
+ * @return bool
+ */
+function jetpack_blaze_post_row_actions_disable( $are_quick_links_enabled, $post ) {
+	if ( 'product' !== $post->post_type ) {
+		return false;
+	}
+
+	return $are_quick_links_enabled;
+}
+add_filter( 'jetpack_blaze_post_row_actions_enable', 'jetpack_blaze_post_row_actions_disable', 10, 2 );