+++ projects/plugins/jetpack/tools/webpack.config.css.js
@@ -178,6 +178,7 @@ for ( const name of [
 	'modules/widget-visibility/widget-conditions/widget-conditions',
 	'modules/widgets/gallery/css/admin',
 	'modules/sso/jetpack-sso-login',
+	'modules/sso/jetpack-sso-admin-create-user',
 	'modules/masterbar/admin-menu/admin-menu',
 	'modules/masterbar/admin-menu/admin-menu-nav-unification',
 ] ) {